requirements.txt
absl-py==0.11.0
addict==2.4.0
aim==2.6.0
aimrecords==0.0.7
alembic==1.6.0
aniso8601==9.0.1
anytree==2.8.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base58==2.0.1
basemap==1.2.0
beautifulsoup4 @ file:///home/linux1/recipes/ci/beautifulsoup4_1610988766420/work
bleach==3.2.1
blobfile==1.2.1
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1598884132938/work
colorama==0.4.4
contrastive-learner==0.1.1
cycler==0.10.0
cytoolz==0.11.0
dask==2.17.0
decorator @ file:///home/ktietz/src/ci/decorator_1611930055503/work
defusedxml==0.6.0
docker==5.0.0
einops==0.3.0
entrypoints==0.3
fastrlock==0.5
filelock==3.0.12
fire==0.4.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==2.7.0
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
GDAL @ file:///E:/pkgs/GDAL-3.2.1-cp37-cp37m-win_amd64.whl
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.24.0
google-auth-oauthlib==0.4.2
greenlet==1.1.0
grpcio==1.34.0
gunicorn==20.1.0
idna==2.10
imageio @ file:///tmp/build/80754af9/imageio_1594161405741/work
importlib-metadata==3.3.0
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.2
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.0.1
kornia==0.5.4
llvmlite==0.24.0
lmdb==1.2.1
lxml==4.2.5
Mako==1.1.4
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
mmcv==1.2.2
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx @ file:///tmp/build/80754af9/networkx_1598376031484/work
notebook==6.1.6
numpy==1.20.1
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.46
osgeo==0.0.0
packaging==20.8
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
Pillow==8.0.1
prefetch-generator==1.0.1
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.14.0
psutil==5.8.0
py==1.10.0
py3nvml==0.2.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodomex==3.10.1
Pygments==2.7.3
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pyproj==1.9.5.1
pyrser==0.2.0
pyrsistent==0.17.3
pyshp @ file:///tmp/build/80754af9/pyshp_1610641829514/work
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
python-editor==1.0.4
pytz==2020.5
PyWavelets==1.1.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
retry==0.9.2
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.24.0
scipy @ file:///C:/ci/scipy_1614023124510/work
seaborn==0.11.1
Send2Trash==1.5.0
six @ file:///home/linux1/recipes/ci/six_1610970791821/work
smmap==4.0.0
soupsieve @ file:///tmp/build/80754af9/soupsieve_1614023247537/work
spark-parser==1.8.9
SQLAlchemy==1.4.13
stylegan2-pytorch==1.8.4
TBB==0.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.12.8
toolz @ file:///home/linux1/recipes/ci/toolz_1610987900194/work
torch==1.7.1+cu101
torchaudio==0.7.2
torchvision==0.8.2+cu101
tornado==5.1.1
tqdm==4.55.0
traitlets==5.0.5
typing-extensions==3.7.4.3
uncompyle==2.0.0
uncompyle6==3.7.4
urllib3==1.26.2
vector-quantize-pytorch==0.1.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
xdis==5.0.6
xlrd==1.2.0
xlwt==1.3.0
xmltodict==0.12.0
yacs==0.1.8
yapf==0.30.0
zipp==3.4.0
